significantly improved extraction
the old approach (compiling object files for the binding and then building up a complicated gcc build command) was brittle, error prone, and system specific. we now compile the binding by replacing the conf tool, which all LKC-based systems ship with. this is a big breaking change which may impact extraction of already supported systems. however, it also greatly improves applicability on other systems, for example toybox now works correctly. the idea is to iron out any remaining problems with the new approach.