improved lambdas and hook functions
previously, lambdas were passed around as source code, and their arguments were interpreted and validated at runtime. now we precompile lambdas into shell scripts, which are then preprocessed and sourced like all the other scripts, and passed around by name only. this is a bit less elegant, but it cuts down the overhead for transformation and solving considerably.

also, hooks are now no longer implemented as lambda functions, because this wasn't used anyway.

... turns out bash doesn't really lend itself well to a functional style ;) who would've thought.

on the upside: lambdas are now profileable.