added fine-grained restriction mechanism to skip specific actions
this is very useful to avoid unnecessary computation effort for nonsensical combinations (e.g., model counting for Plaisted-Greenbaum transformation, distributive transformation on large models, or model counting on Linux versions known to be too complex). also this can be used to control better what happens in which pass of a large and complex multi-pass experiment.